Several UI fixes to VPN site troubleshooting view
