+++ LocalPackages/NetworkProtectionMac/Sources/NetworkProtectionUI/Extensions/View+NetworkProtection.swift
@@ -1,0 +1,102 @@
+//
+//  View+NetworkProtection.swift
+//
+//  Copyright © 2024 DuckDuckGo. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import SwiftUI
+
+private enum Opacity {
+    static func connectionStatusDetail(colorScheme: ColorScheme) -> Double {
+        colorScheme == .light ? Double(0.6) : Double(0.5)
+    }
+
+    static func dataVolume(colorScheme: ColorScheme) -> Double {
+        colorScheme == .light ? Double(0.6) : Double(0.5)
+    }
+
+    static let content = Double(0.58)
+    static let label = Double(0.9)
+    static let link = Double(1)
+
+    static func sectionHeader(colorScheme: ColorScheme) -> Double {
+        colorScheme == .light ? Double(0.84) : Double(0.85)
+    }
+
+    static func timer(colorScheme: ColorScheme) -> Double {
+        colorScheme == .light ? Double(0.6) : Double(0.5)
+    }
+
+    static func title(colorScheme: ColorScheme) -> Double {
+        colorScheme == .light ? Double(0.84) : Double(0.85)
+    }
+}
+
+extension View {
+    func applyConnectionStatusDetailAttributes(colorScheme: ColorScheme) -> some View {
+        opacity(Opacity.connectionStatusDetail(colorScheme: colorScheme))
+            .font(.NetworkProtection.connectionStatusDetail)
+            .foregroundColor(Color(.defaultText))
+    }
+
+    func applyDataVolumeAttributes(colorScheme: ColorScheme) -> some View {
+        opacity(Opacity.dataVolume(colorScheme: colorScheme))
+            .font(.NetworkProtection.dataVolume)
+            .foregroundColor(Color(.defaultText))
+    }
+
+    func applyCurrentSiteAttributes() -> some View {
+        font(.NetworkProtection.currentSite)
+    }
+
+    func applyLocationAttributes() -> some View {
+        font(.NetworkProtection.location)
+    }
+
+    func applyContentAttributes(colorScheme: ColorScheme) -> some View {
+        opacity(Opacity.content)
+            .font(.NetworkProtection.content)
+            .foregroundColor(Color(.defaultText))
+    }
+
+    func applyDescriptionAttributes() -> some View {
+        font(.NetworkProtection.description)
+            .foregroundColor(Color(.secondaryText))
+    }
+
+    func applyLabelAttributes(colorScheme: ColorScheme) -> some View {
+        opacity(Opacity.label)
+            .font(.NetworkProtection.label)
+            .foregroundColor(Color(.defaultText))
+    }
+
+    func applySectionHeaderAttributes(colorScheme: ColorScheme) -> some View {
+        opacity(Opacity.sectionHeader(colorScheme: colorScheme))
+            .font(.NetworkProtection.sectionHeader)
+            .foregroundColor(Color(.defaultText))
+    }
+
+    func applyTimerAttributes(colorScheme: ColorScheme) -> some View {
+        opacity(Opacity.timer(colorScheme: colorScheme))
+            .font(.NetworkProtection.timer)
+            .foregroundColor(Color(.defaultText))
+    }
+
+    func applyTitleAttributes(colorScheme: ColorScheme) -> some View {
+        opacity(Opacity.title(colorScheme: colorScheme))
+            .font(.NetworkProtection.title)
+            .foregroundColor(Color(.defaultText))
+    }
+}
